// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  ADMIN: 'ADMIN',
  MEMBER: 'MEMBER'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const ProjectType = {
  NOVA_CONSTRUCAO: 'NOVA_CONSTRUCAO',
  REFORMA: 'REFORMA',
  AMPLIACAO: 'AMPLIACAO'
} as const

export type ProjectType = (typeof ProjectType)[keyof typeof ProjectType]


export const ProjectStatus = {
  RASCUNHO: 'RASCUNHO',
  FINALIZADO: 'FINALIZADO',
  ENVIADO: 'ENVIADO'
} as const

export type ProjectStatus = (typeof ProjectStatus)[keyof typeof ProjectStatus]


export const PhaseType = {
  TERRAPLENAGEM: 'TERRAPLENAGEM',
  FUNDACAO: 'FUNDACAO',
  ESTRUTURA_ALVENARIA: 'ESTRUTURA_ALVENARIA',
  LAJE: 'LAJE',
  INSTALACOES_ELETRICAS: 'INSTALACOES_ELETRICAS',
  INSTALACOES_HIDROSSANITARIAS: 'INSTALACOES_HIDROSSANITARIAS',
  ESCADA: 'ESCADA',
  REVESTIMENTOS: 'REVESTIMENTOS',
  PINTURA: 'PINTURA',
  COBERTURA: 'COBERTURA',
  ACABAMENTO: 'ACABAMENTO',
  OUTROS: 'OUTROS'
} as const

export type PhaseType = (typeof PhaseType)[keyof typeof PhaseType]


export const LaborModel = {
  FIXED: 'FIXED',
  PER_M2: 'PER_M2',
  PERCENT: 'PERCENT'
} as const

export type LaborModel = (typeof LaborModel)[keyof typeof LaborModel]


export const MaterialCategory = {
  TERRAPLENAGEM: 'TERRAPLENAGEM',
  FUNDACAO: 'FUNDACAO',
  LAJE: 'LAJE',
  PINTURA: 'PINTURA',
  ACABAMENTO: 'ACABAMENTO',
  ESTRUTURA: 'ESTRUTURA',
  ALVENARIA: 'ALVENARIA',
  COBERTURA: 'COBERTURA',
  ELETRICA: 'ELETRICA',
  HIDRAULICA: 'HIDRAULICA',
  REVESTIMENTO: 'REVESTIMENTO',
  ESQUADRIA: 'ESQUADRIA',
  OUTROS: 'OUTROS'
} as const

export type MaterialCategory = (typeof MaterialCategory)[keyof typeof MaterialCategory]